+++ src/main/server/server-online-map/server.ts
@@ -2,8 +2,7 @@ import axios from 'axios';
 import { ENGINE_PATH } from '../../constant';
 import express from 'express';
 import path from 'path';
-import getWorldSettingsByServerId from '../../services/worldSettings/getWorldSettingsByServerId';
-import trimWorldSettingsString from '../../../utils/trimWorldSettingsString';
+import getAdminConnectionConfig from '../../services/admin/getAdminHost';
 
 const app = express();
 const port = 3434;
@@ -13,14 +12,11 @@ app.use(express.static(path.join(ENGINE_PATH, 'server-online-map', 'public')));
 app.get('/:serverId/players', async (req, res) => {
   const serverId = req.params.serverId;
 
-  const worldSettings = await getWorldSettingsByServerId(serverId);
-
-  const ipAdress = '127.0.0.1';
-  const apiPort = worldSettings.RESTAPIPort || 8212;
-  const adminPassword = trimWorldSettingsString(worldSettings.AdminPassword);
+  const { host, restPort, adminPassword } =
+    await getAdminConnectionConfig(serverId);
 
   try {
-    const result = await axios(`http://${ipAdress}:${apiPort}/v1/api/players`, {
+    const result = await axios(`http://${host}:${restPort}/v1/api/players`, {
       method: 'get',
       auth: {
         username: 'admin',
@@ -37,14 +33,11 @@ app.get('/:serverId/players', async (req, res) => {
 app.get('/:serverId/info', async (req, res) => {
   const serverId = req.params.serverId;
 
-  const worldSettings = await getWorldSettingsByServerId(serverId);
-
-  const ipAdress = '127.0.0.1';
-  const apiPort = worldSettings.RESTAPIPort || 8212;
-  const adminPassword = trimWorldSettingsString(worldSettings.AdminPassword);
+  const { host, restPort, adminPassword } =
+    await getAdminConnectionConfig(serverId);
 
   try {
-    const result = await axios(`http://${ipAdress}:${apiPort}/v1/api/info`, {
+    const result = await axios(`http://${host}:${restPort}/v1/api/info`, {
       method: 'get',
       auth: {
         username: 'admin',