+++ src/types/RemoteSettings.types.ts
@@ -1,0 +1,9 @@
+export type RemoteSettings = {
+  ServerName: string;
+  PublicIP: string;
+  RESTAPIPort: number;
+  RCONPort: number;
+  AdminPassword: string;
+  RESTAPIEnabled: boolean;
+  RCONEnabled: boolean;
+};